% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict_AML_clusters}
\alias{predict_AML_clusters}
\title{Transcriptional AML cluster prediction}
\usage{
predict_AML_clusters(matrix)
}
\arguments{
\item{matrix}{Sample x genes matrix}
}
\value{
A dataframe with for each of the 17 clusters the distance to the decision boundary, the cluster prediction and if the distance to the boundary was higher than the cutoff. Rownames of the imput matrix will be set as rownames and under the column "sample_id".
}
\description{
Predict the clusters found in "ref paper".
}
